clients/requirements.txt
google-cloud-speech>=2.27.0
google-cloud-pubsub>=2.21.0
google-cloud-texttospeech>=2.16.5
sounddevice>=0.4.6
numpy>=1.26.0
requests>=2.32.0
